feat: Creating OverviewSpendBar
