+++ lib/invoices/screens/Invoices.dart
@@ -6,6 +6,7 @@ import 'package:finance_controlinator_mobile/invoices/services/SyncStorageServic
 
 import 'package:flutter/material.dart';
 import '../../expenses/components/OverviewCard.dart';
+import '../../expenses/components/OverviewSpendBar.dart';
 import '../domain/sync/InvoiceSync.dart';
 import '../webclients/InvoiceSyncWebClient.dart';
 
@@ -36,9 +37,7 @@ class InvoicesScreen extends StatelessWidget {
             appBar: AppBar(
               title: Text(response.syncName), //from dto
             ),
-            body: Column(
-              children: [InvoiceMonthDatasScreen(response.monthDataSyncs)],
-            ),
+            body: InvoiceMonthDatasScreen(response.monthDataSyncs),
           );
         }
         // },
@@ -105,7 +104,10 @@ class InvoiceOverviewScreen extends StatelessWidget {
                     child: OverviewBriefs(overview.briefs)),
                 Padding(
                     padding: const EdgeInsets.only(bottom: 8),
-                    child: OverviewPartitions(overview.partitions)),
+                    child: OverviewSpendBar(overview.partitions
+                        .map((e) => PartitionData(
+                            e.type, e.typeText, e.percent, e.totalValue))
+                        .toList())),
               ],
             )));
   }
@@ -148,104 +150,6 @@ class OverviewBriefs extends StatelessWidget {
         children: briefs.map((e) => OverviewCard(e.text)).toList(),
       ),
     );
-  }
-}
-
-class OverviewPartitions extends StatelessWidget {
-  List<InvoicePartition> partitions;
-
-  OverviewPartitions(this.partitions, {Key? key}) : super(key: key);
-
-  List<Color> colors = [
-    Colors.redAccent,
-    Colors.greenAccent,
-    Colors.blueAccent,
-    Colors.orangeAccent,
-    Colors.blueGrey,
-    Colors.purpleAccent,
-    Colors.deepOrangeAccent,
-  ];
-
-  @override
-  Widget build(BuildContext context) {
-    var descriptions = buildDescriptions(partitions, colors);
-
-    return Column(
-      children: [
-        Padding(
-            padding: const EdgeInsets.only(bottom: 8),
-            child: buildBar(partitions, colors)),
-        Padding(
-            padding: const EdgeInsets.only(bottom: 8),
-            child: Row(
-                mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                children: descriptions.take(4).toList())),
-        Row(
-            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-            children: descriptions.skip(4).take(4).toList())
-      ],
-    );
-  }
-
-  List<Row> buildDescriptions(
-      List<InvoicePartition> parts, List<Color> colors) {
-    return parts
-        .asMap()
-        .map((key, value) => MapEntry(
-            key,
-            Row(
-              children: [
-                Padding(
-                  padding: const EdgeInsets.only(left: 4, right: 4),
-                  child: Text(value.typeText + ":"),
-                ),
-                Container(
-                    height: 20,
-                    width: 20,
-                    decoration: BoxDecoration(
-                        color: colors[key],
-                        borderRadius:
-                            const BorderRadius.all(Radius.circular(100))))
-              ],
-            )))
-        .values
-        .toList();
-  }
-
-//todo: make a component
-  Row buildBar(List<InvoicePartition> parts, List<Color> colors) {
-    var partitionsWithSpend =
-        parts.where((element) => element.totalValue > 0).toList();
-
-    return Row(
-        children: partitionsWithSpend
-            .asMap()
-            .map((key, value) {
-              return MapEntry(
-                  key,
-                  Expanded(
-                      flex: value.percent.toInt(),
-                      child: Container(
-                        height: 20,
-                        decoration: BoxDecoration(
-                            color: colors[key],
-                            borderRadius:
-                                getBorder(key, partitionsWithSpend.length - 1)),
-                      )));
-            })
-            .values
-            .toList());
-  }
-
-//todo: be part of the component
-  BorderRadius getBorder(int itemPosition, int lastPosition) {
-    return itemPosition == 0
-        ? const BorderRadius.only(
-            topLeft: Radius.circular(6), bottomLeft: Radius.circular(6))
-        : itemPosition == lastPosition
-            ? const BorderRadius.only(
-                topRight: Radius.circular(6), bottomRight: Radius.circular(6))
-            : const BorderRadius.all(Radius.zero);
   }
 }
